// src/app/views/base/model/auto.model.ts
export class AutoModel {
    _id: string= '';
    marca: string = '';
    modelo: string = '';
    year: number = 0;
    color: string = '';
    tipocombustible: string = '';
    capacidad: string = '';
    transmision: string = '';
    precio: string = '';
    precioventa: number = 0;
    disponibilidad: string = '';
    descripcion: string = '';
    prioridad: number = 0;
    mensaje: string | undefined;
}